services refactor: post and todo
