+++ src/services/Post/getPosts.ts
@@ -1,5 +1,5 @@
-import { Post } from "../constants/post";
-import { api } from "./api";
+import { Post } from "../../constants/post";
+import { api } from "../api";
 
 export async function getPosts(): Promise<Post[]> {
   const { data } = await api.get("/posts");
