+++ src/services/Todo/addTodo.ts
@@ -1,5 +1,5 @@
-import { newTodo } from "../constants/todo";
-import { api } from "./api";
+import { newTodo } from "../../constants/todo";
+import { api } from "../api";
 
 export async function addTodo(body: newTodo): Promise<void> {
   await api.post("/todo", body);
